1st submit of design tool peer-reviewed
This is the 9th version of a local tool. After request for sharing, I am pushing it online. 
This tool has been created in an open source spirit with the objective to support the design of small solar systems.
It has been applied to several installation in Uganda and has been peer-reviewed by external consultant.
A tentative Bills of Quantities is estimated for the case of Uganda, but it is market specific.